Remove attributes annotation of Builder `struct`
The `struct` created for a Builder should not have attributes because these structs only derive from `Debug`,
`PartialEq`, `Clone` and `Default`. None of these support attributes.

This becomes a problem if a plugin tries to add attributes to the `metadata` field to configure the generated code for
the `struct`. In this case, the attributes will also be added to the `Builder` struct; which is wrong.