add io related abstractions to support s3 tm uploads (#3756)
## Description

Adds a new `io` module to S3 transfer manager HLL. This is supporting
functionality for implementing uploads.

High level description of new components:

* **InputStream** - this is a new abstraction to replace `ByteStream`
for use in transfer manager uploads. The primary driver behind this is
once we create a `ByteStream` we have very little control over
optimizing how it can be consumed. We lose whether it came from a file
or in-memory or elsewhere. Internally this will of course eventually map
to a `ByteStream` when we call `UploadPart` or `PutObject` but we want
to have control over how bytes are read from the stream (e.g. being able
to read a file concurrently).
* In the future we may add additional variants to support e.g. unknown
streams (e.g. `impl io::Read`) or giving users control over checksums
for each part, etc. This is an area of exploration still.
* **PartReader / ReadPart** - internal abstractions over `InputStream`
that workers will eventually consume. The idea is we know what
`InputStream` is and we convert it to something that reads "parts" for
multipart upload in the most efficient way possible for whatever the
underlying stream type is.
* In a subsequent PR workers share a single part reader and read from it
concurrently. Workers don't have to care about anything other than
asking for the next "part" of data and uploading it.
* **PathBodyBuilder / PathBody** - replacements for `FsBuilder` from
smithy runtime / SDK. Notable difference is we don't currently allow
specifying either an already open file or control over buffer sizes.
* We can of course support an open file but the implementation will
require additional locking to ensure bytes are read sequentially

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._