Remove unused hyper dependency in aws-config (#3770)
This allows end-users to keep hyper 0.x out of their tree

## Motivation and Context
This dependency appears unused and means that I have a 0.14 import in my
tree

## Description
Removes the dependency

## Testing
Builds locally, if CI passes I think its fair to say this is a no-op. I
can't find any usages of this lib.

## Checklist
I'm not sure this change warrants a changelog as there is no user facing
change

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._