Fix AWS SDK generation examples in  `aws/sdk` README (#3862)
## Motivation and Context
#3852 

## Testing
Manually executed the updated examples and confirmed they passed on my
mac.

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [x] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.

Even though the fix is for AWS SDK code generation examples, I think the
release notes for the fix should only appear in the `smithy-rs`
repository. Customers subscribing to the `aws-sdk-rust` repository
should not need to know.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._