chore: add Into conversions for Blob to [u8] (#3877)
## Motivation and Context
Often I have a &[u8] or a Vec<u8>, but the sdk function takes a Blob.
It would be convenient to pass the my_vec instead of
aws_smithy_types::Blob::new(my_vec)

## Description
In blob.rs I added
Into<Vec<u8>> for Blob
Into<Blob> for Vec<u8>
Into<Blob> for &[u8]


## Testing
Added a test to blob.rs to test the various scenarios.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._